Fix another four ftests via getPhysicalPath
