+++ addons/pkg-maven/ftests/src/main/java/org/commonjava/indy/pkg/maven/content/GroupMetadataFileDeletionTest.java
@@ -19,6 +19,7 @@ import org.commonjava.indy.client.core.IndyClientException;
 import org.commonjava.indy.ftest.core.AbstractContentManagementTest;
 import org.commonjava.indy.model.core.Group;
 import org.commonjava.indy.model.core.HostedRepository;
+import org.commonjava.indy.util.LocationUtils;
 import org.junit.Before;
 import org.junit.Test;
 
@@ -55,7 +56,7 @@ import static org.jgroups.util.Util.assertTrue;
 public class GroupMetadataFileDeletionTest
                 extends AbstractContentManagementTest
 {
-    private static final String METADATA_PATH = "/org/foo/bar/maven-metadata.xml";
+    private static final String METADATA_PATH = "org/foo/bar/maven-metadata.xml";
 
     /* @formatter:off */
     private static final String METADATA_CONTENT = "<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n" +
@@ -96,8 +97,8 @@ public class GroupMetadataFileDeletionTest
     @Test
     public void run() throws Exception
     {
-        File f1 = new File( storageDir, "maven/group-" + g1.getName() + METADATA_PATH );
-        File f2 = new File( storageDir, "maven/group-" + g2.getName() + METADATA_PATH );
+        File f1 = getPhysicalStorageFile( LocationUtils.toLocation( g1 ), METADATA_PATH );
+        File f2 = getPhysicalStorageFile( LocationUtils.toLocation( g2 ), METADATA_PATH );
 
         // Files exist
         assertTrue( f1.exists() );
@@ -105,6 +106,8 @@ public class GroupMetadataFileDeletionTest
 
         // Delete cache file from G1, which will delete parent G2's cache too
         client.content().deleteCache( g1.getKey(), METADATA_PATH );
+
+        sleepAndRunFileGC( 1000 );
 
         // Verify files were deleted
         assertFalse( f1.exists() );
