+++ ftests/core/src/main/java/org/commonjava/indy/ftest/core/content/GroupMetadataMergeInfoGenTest.java
@@ -24,6 +24,7 @@ import org.commonjava.indy.model.core.Group;
 import org.commonjava.indy.model.core.HostedRepository;
 
 import org.commonjava.indy.model.core.StoreType;
+import org.commonjava.indy.util.LocationUtils;
 import org.junit.Test;
 
 import java.io.ByteArrayInputStream;
@@ -195,12 +196,7 @@ public class GroupMetadataMergeInfoGenTest
     private void assertInfoContent( final ArtifactStore store, final String path, final String expectedContent )
             throws Exception
     {
-
-//        final String infoFilePath =
-//                String.format( "%s/var/lib/indy/storage/%s-%s/%s", fixture.getBootOptions().getIndyHome(), group.name(),
-//                               store.getName(), path + GroupMergeHelper.MERGEINFO_SUFFIX );
-        final File infoFile = Paths.get( fixture.getBootOptions().getHomeDir(), "var/lib/indy/storage", store.getPackageType(),
-                                         group.singularEndpointName() + "-" + store.getName(), path + GroupMergeHelper.MERGEINFO_SUFFIX ).toFile();
+        final File infoFile = getPhysicalStorageFile( LocationUtils.toLocation( store ), path + GroupMergeHelper.MERGEINFO_SUFFIX );
         assertThat( "info file doesn't exist", infoFile.exists(), equalTo( true ) );
 
         try (final InputStream stream = new FileInputStream( infoFile ))
